requirements.txt
astroid==1.5.3
colorama==0.3.9
Django==1.11.3
djangorestframework==3.6.3
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.2
pytz==2017.2
six==1.10.0
wrapt==1.10.10